app/backend/velarion/requirements.txt
Django==4.2
djangorestframework==3.14
drf-spectacular==0.28
psycopg2-binary
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
lightgbm
joblib==1.4.2
celery==5.5.3
flower==2.0.1
requests==2.32.3
redis>=4.5.0